Move common types out of oneCursor
